+++ src/Data/Constraint/If.hs
@@ -1,28 +1,42 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneKindSignatures #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE UndecidableInstances #-}
 
 {-|
 Module: Data.Constraint.If
 
-This module defines the typeclass 'IfSat', with method 'ifSat':
+This module defines the constraint disjunction typeclass '||', with method 'dispatch':
 
-> ifSat :: forall (ct :: Constraint) (r :: Type). IfSat ct => ( ct => r ) -> r -> r
+> dispatch :: ( c || d ) => ( c => r ) -> ( d => r ) -> r
 
-An expression of the form @ ifSat \@ct yes no @ denotes a selection between the two
+An expression of the form @ dispatch \@c \@d yes no @ denotes a selection between the two
 branches @yes@ and @no@:
 
-  - if the constraint @ct@ can be determined to hold at the point of solving @IfSat ct@,
-    then the @yes@ branch is selected, which has access to the @ct@ constraint;
-  - otherwise, the fallback branch @no@ is selected.
+  - if the constraint @c@ can be determined to hold at the point of solving @c || d@,
+    then the @yes@ branch is selected, which has access to evidence for the @c@ constraint;
+  - otherwise, the fallback branch @no@ is selected, which has access to evidence for @d@.
+
+If you don't need additional constraints in the fallback branch, you can also use:
+
+> ifSat :: IfSat c => ( c => r ) -> r -> r
+
+This is the special case of 'dispatch' which taes @d@ to be the trivial constraint,
+@d ~ ( () :: Constraint)@.
 
 This module also provides the type family @'IsSat' :: Constraint -> Bool@, which, when reduced,
 will check whether the constraint provided as an argument is satisfied.
 
-To use this module, you will also need to enable the corresponding 'IfSat.Plugin.plugin',
+To use this functionality, you must enable the corresponding 'IfSat.Plugin.plugin',
 by adding @\{\-\# OPTIONS_GHC -fplugin=IfSat.Plugin \#\-\}@
 to the header of your module.
 
@@ -112,7 +126,7 @@ different contexts.
 -}
 
 module Data.Constraint.If
-  ( IfSat(..), IsSat )
+  ( type (||)(dispatch), IfSat, ifSat, IsSat )
   where
 
 -- base
@@ -121,23 +135,52 @@ import Data.Kind
 
 --------------------------------------------------------------------------------
 
-type IfSat :: Constraint -> Constraint
-class IfSat ct where
-  -- | @ IfSat \@ct a b@ returns @a@ if the constraint is satisfied,
-  -- and @b@ otherwise.
+
+type (||) :: Constraint -> Constraint -> Constraint
+class c || d where
+  -- | @dispatch \@c \@d a b@ returns @a@ if the constraint @c@ is satisfied,
+  -- otherwise @b@.
+  --
+  -- > dispatch :: ( c || d ) => ( c => r ) -> ( d => r ) -> r
   --
   -- Requires the if-instance 'IfSat.Plugin.plugin':
   -- add @{-# OPTIONS_GHC -fplugin=IfSat.Plugin #-}@
   -- to the header of your module.
   --
-  -- Note: the selection happens at the point in the code where the @IfSat ct@
+  -- Note: the selection happens at the point in the code where the @c || d@
   -- constraint is solved.
-  ifSat :: ( ( IsSat ct ~ True, ct ) => r )
-        -> ( IsSat ct ~ False => r )
-        -> r
+  dispatch :: ( ( IsSat c ~ True, c ) => r )
+           -> ( ( IsSat c ~ False, IsSat d ~ True, d ) => r )
+           -> r
+
+type IfSat :: Constraint -> Constraint
+type IfSat ct = ( ct || () )
+  -- IfSat must be a type synonym, not a class newtype such as
+  --
+  -- > class    ( ct || () ) => IfSat ct
+  -- > instance ( ct || () ) => IfSat ct
+  --
+  -- Otherwise, mentions of 'IfSat' can cause GHC to rewrite
+  -- too early.
+
+-- | @ ifSat \@ct a b@ returns @a@ if the constraint @ct@ is satisfied,
+-- and @b@ otherwise.
+--
+-- Requires the if-instance 'IfSat.Plugin.plugin':
+-- add @{-# OPTIONS_GHC -fplugin=IfSat.Plugin #-}@
+-- to the header of your module.
+--
+-- Note: the selection happens at the point in the code where the @IfSat ct@
+-- constraint is solved.
+ifSat :: forall ct r
+      .  ( IfSat ct )
+      => ( ( IsSat ct ~ True, ct ) => r )
+      -> ( ( IsSat ct ~ False ) => r )
+      -> r
+ifSat f g = dispatch @ct @() f g
 
 -- | @IsSat ct@ returns @True@ if @ct@ is satisfied, and @False@ otherwise.
 --
--- The satisfiability check occurs at the moment of type-family reduction.
+-- The satisfiability check occurs at the moment of performing type-family reduction.
 type IsSat :: Constraint -> Bool
 type family IsSat ct where