+++ changelog.md
@@ -1,4 +1,20 @@
 
+# Version 0.3.0.0 (2021-09-01)
+
+- Add the `(||)` constraint disjunction mechanism, with
+
+```haskell
+dispatch :: ( c || d ) => ( c => r ) -> ( d => r ) -> r
+```
+
+This allows users to select between two different constraints.
+This is more general than `ifCt`, as it allows extra constraints
+in the fallback branch.
+
+- Implement `IfCt`, `ifCt` in terms of `(||)` and `dispatch`.
+  `IfCt` becomes a type synonym, which means that it no longer
+  bundles `ifCt`.
+
 # Version 0.2.1.1 (2021-08-31)
 
 - Minor documentation improvements.